+++ NEWS.md
@@ -1,5 +1,8 @@
 # S7 (development version)
 
+* In `new_class()`, properties can either be named by naming the element
+  of the list or by supplying the `name` argument to `new_property()` (#371).
+
 * `super()` now works with Ops methods (#357).
 
 * `method()` now generates an informative message when dispatch fails (#387).